✨ refactor: 스도쿠 셀 관련 유틸리티 함수 제거 및 코드 최적화 - 셀 스타일 관련 로직을 훅으로 통합하고, 불필요한 테스트 코드 정리
