+++ src/features/game-board/model/hooks/index.ts
@@ -1,1 +1,2 @@
+export * from "./useCell";
 export * from "./useKillerCage";
